+++ fake-cucumber/javascript/src/CucumberStream.ts
@@ -2,11 +2,15 @@ import { Transform, TransformCallback } from 'stream'
 import { messages } from 'cucumber-messages'
 import Cucumber from './Cucumber'
 import IStepDefinition from './IStepDefinition'
+import { IHook } from './IHook'
 
 export default class CucumberStream extends Transform {
   private readonly gherkinMessages: messages.IEnvelope[] = []
 
-  constructor(private readonly stepDefinitions: IStepDefinition[]) {
+  constructor(
+    private readonly stepDefinitions: IStepDefinition[],
+    private readonly hooks: IHook[]
+  ) {
     super({ objectMode: true })
   }
 
@@ -20,7 +24,11 @@ export default class CucumberStream extends Transform {
   }
 
   public _flush(callback: TransformCallback): void {
-    const cucumber = new Cucumber(this.gherkinMessages, this.stepDefinitions)
+    const cucumber = new Cucumber(
+      this.gherkinMessages,
+      this.stepDefinitions,
+      this.hooks
+    )
     cucumber.execute(message => this.push(message))
     callback()
   }