+++ fake-cucumber/javascript/src/ExpressionStepDefinition.ts
@@ -8,13 +8,14 @@ import {
 import { messages } from 'cucumber-messages'
 import uuidv4 from 'uuid/v4'
 import IStepDefinition from './IStepDefinition'
+import { AnyBody } from './types'
 
 export default class ExpressionStepDefinition implements IStepDefinition {
   private readonly id = uuidv4()
 
   constructor(
     private readonly expression: Expression,
-    private readonly body: (...args: any) => any
+    private readonly body: AnyBody
   ) {}
 
   public match(